+++ lib/puzzletaskHelpers.ts
